requirements.txt
requests
tweepy
python-dotenv
google-genai